\name{gLRTH-package}
\alias{gLRTH-package}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
A Likelihood Ratio Test for Genome-Wide Association under Genetic Heterogeneity
}
\description{
Functions for the Likelihood Ratio Test for Genome-Wide Association under Genetic Heterogeneity
}

\references{
Qian, M., & Shao, Y. (2013). A Likelihood Ratio Test for Genome-Wide Association under Genetic Heterogeneity. Annals of human genetics, 77(2), 174-182.
}
\author{
Xiaoxia Han and Yongzhao Shao
}


